+++ vmmlib/vector.hpp
@@ -210,23 +210,14 @@ public:
     // compute the cross product of two vectors
     // note: there's also a free function:
     // vector<> cross( const vector<>, const vector<> )
-
-    // result = vec1.cross( vec2 ) => retval result = vec1 x vec2
     template< typename TT >
-    vector cross( const vector< M, TT >& rhs,
-                  typename enable_if< M == 3, TT >::type* = 0 ) const;
-
-    // result.cross( vec1, vec2 ) => (this) = vec1 x vec2
-    template< typename TT >
-    void cross( const vector< M, TT >& a, const vector< M, TT >& b,
-                typename enable_if< M == 3, TT >::type* = 0 );
-
+    vector< M, T >& cross( const vector< M, TT >& b,
+                           typename enable_if< M == 3, TT >::type* = 0 );
 
     // compute the dot product of two vectors
     // note: there's also a free function:
     // T dot( const vector<>, const vector<> );
     inline T dot( const vector& other ) const;
-
 
     // normalize the vector
     // note: there's also a free function:
@@ -248,25 +239,20 @@ public:
     /** @return the product of all elements of this vector */
     T product() const;
 
+    /** Rotate this 3-component vector around the given axis. */
     template< typename TT >
-    vector< 3, T > rotate( const T theta, vector< M, TT > axis,
-                           typename enable_if< M == 3, TT >::type* = 0 ) const;
+    vector< M, T >& rotate( T theta, vector< M, TT > axis,
+                            typename enable_if< M == 3, TT >::type* = 0 );
 
-    // right hand system, CCW triangle
-    // (*this) = normal of v0,v1,v2
-    void compute_normal( const vector& v0, const vector& v1, const vector& v2 );
-    // retval = normal of (this), v1, v2
-    vector compute_normal( const vector& v1, const vector& v2 ) const;
+    /** @return the sub vector of the given length at the given offset. */
+    template< size_t N, size_t O = 0 >
+    vector< N, T > get_sub_vector( typename enable_if< M >= N+O >::type* = 0 )
+        const;
 
-    /** @return the sub vector at the given position and length. */
-    template< size_t N >
-    vector< N, T >& get_sub_vector( size_t offset = 0,
-        typename enable_if< M >= N >::type* = 0 );
-
-    /** @return the sub vector at the given position and length. */
-    template< size_t N >
-    const vector< N, T >& get_sub_vector( size_t offset = 0,
-        typename enable_if< M >= N >::type* = 0 ) const;
+    /** Set the sub vector of the given length at the given offset. */
+    template< size_t N, size_t O = 0 >
+    void set_sub_vector( const vector< N, T >& sub,
+                         typename enable_if< M >= N+O >::type* = 0 );
 
     // sphere functions - sphere layout: center xyz, radius w
     template< typename TT >
@@ -304,10 +290,6 @@ public:
     const T&    find_max() const;
 
     void clamp( const T& min = 0.0, const T& max = 1.0 );
-
-    template< typename TT >
-    void scale_to( vector< M, TT >& scaled_vector,
-        T min_value = -1.0, T max_value = 1.0 ) const;
 
     inline static size_t size(); // returns M
 
@@ -431,7 +413,6 @@ bool equals( const vector< M, T >& a, const vector< M, T >& b )
     return a.equals( b );
 }
 
-
 // allows float * vector, not only vector * float
 template< size_t M, typename T >
 static vector< M, T > operator* ( T factor, const vector< M, T >& vector_ )
@@ -439,18 +420,34 @@ static vector< M, T > operator* ( T factor, const vector< M, T >& vector_ )
     return vector_ * factor;
 }
 
-
 template< size_t M, typename T >
 inline T dot( const vector< M, T >& first, const vector< M, T >& second )
 {
     return first.dot( second );
 }
 
-
 template< size_t M, typename T >
-inline vector< M, T > cross( const vector< 3, T >& a, const vector< 3, T >& b )
+inline vector< M, T > cross( vector< M, T > a, const vector< M, T >& b )
 {
     return a.cross( b );
+}
+
+template< size_t M, typename T >
+vector< M, T > compute_normal( const vector< M, T >& a, const vector< M, T >& b,
+                               const vector< M, T >& c )
+{
+    // right hand system, CCW triangle
+    const vector< M, T > u = b - a;
+    const vector< M, T > v = c - a;
+    vector< M, T > w = cross( u, v );
+    w.normalize();
+    return w;
+}
+
+template< typename T >
+vector< 3, T > rotate( const vector< 3, T >& vec, T theta, vector< 3, T > axis )
+{
+    return vector< 3, T >( vec ).rotate( theta, axis );
 }
 
 
@@ -470,7 +467,7 @@ inline vector< 4, T > compute_plane( const vector< 3, T >& a,
     const vector< 3, T > cb = b - c;
     const vector< 3, T > ba = a - b;
 
-    vector< 4, T > plane = cb.cross( ba );
+    vector< 4, T > plane = vector< 4, T >( cross( cb, ba ));
     plane.normalize();
     plane.w() = -plane.x() * a.x() - plane.y() * a.y() - plane.z() * a.z();
     return plane;
@@ -492,7 +489,6 @@ vector< M, T >::vector( const T& _x, const T& _y )
     array[ 1 ] = _y;
 }
 
-
 template< size_t M, typename T >
 vector< M, T >::vector( const T& _x, const T& _y, const T& _z )
 {
@@ -500,8 +496,6 @@ vector< M, T >::vector( const T& _x, const T& _y, const T& _z )
     array[ 1 ] = _y;
     array[ 2 ] = _z;
 }
-
-
 
 template< size_t M, typename T >
 vector< M, T >::vector( const T& _x, const T& _y, const T& _z, const T& _w )
@@ -511,7 +505,6 @@ vector< M, T >::vector( const T& _x, const T& _y, const T& _z, const T& _w )
     array[ 2 ] = _z;
     array[ 3 ] = _w;
 }
-
 
 template< size_t M, typename T >
 vector< M, T >::vector( const T* values )
@@ -549,8 +542,6 @@ vector< M, T >::vector( const vector< M-1, T >& vector_, T last_ )
 }
 #endif
 
-
-
 // to-homogenous-coordinates ctor
 template< size_t M, typename T >
 template< size_t N >
@@ -559,9 +550,6 @@ vector< M, T >::vector( const vector< N, T >& source_,
 {
     (*this) = source_;
 }
-
-
-
 
 // from-homogenous-coordinates ctor
 template< size_t M, typename T >
@@ -572,7 +560,6 @@ vector< M, T >::vector( const vector< N, T >& source_,
     (*this) = source_;
 }
 
-
 template< size_t M, typename T >
 template< typename U >
 vector< M, T >::vector( const vector< M, U >& source_ )
@@ -580,14 +567,11 @@ vector< M, T >::vector( const vector< M, U >& source_ )
     (*this) = source_;
 }
 
-
-
 template< size_t M, typename T > void vector< M, T >::set( T _a )
 {
     for( iterator it = begin(), it_end = end(); it != it_end; ++it )
         *it = _a;
 }
-
 
 #ifndef SWIG
 template< size_t M, typename T >
@@ -613,7 +597,6 @@ void vector< M, T >::set( T _x, T _y )
     array[ 1 ] = _y;
 }
 
-
 template< size_t M, typename T >
 void vector< M, T >::set( T _x, T _y, T _z )
 {
@@ -621,8 +604,6 @@ void vector< M, T >::set( T _x, T _y, T _z )
     array[ 1 ] = _y;
     array[ 2 ] = _z;
 }
-
-
 
 template< size_t M, typename T >
 void vector< M, T >::set( T _x, T _y, T _z, T _w )
@@ -633,7 +614,6 @@ void vector< M, T >::set( T _x, T _y, T _z, T _w )
     array[ 3 ] = _w;
 }
 
-
 template< size_t M, typename T >
 inline T&
 vector< M, T >::operator()( size_t index )
@@ -641,16 +621,12 @@ vector< M, T >::operator()( size_t index )
     return at( index );
 }
 
-
-
 template< size_t M, typename T >
 inline const T&
 vector< M, T >::operator()( size_t index ) const
 {
     return at( index );
 }
-
-
 
 template< size_t M, typename T >
 inline T&
@@ -665,8 +641,6 @@ vector< M, T >::at( size_t index )
     return array[ index ];
 }
 
-
-
 template< size_t M, typename T >
 inline const T&
 vector< M, T >::at( size_t index ) const
@@ -680,7 +654,6 @@ vector< M, T >::at( size_t index ) const
     return array[ index ];
 }
 
-
 #ifndef VMMLIB_NO_CONVERSION_OPERATORS
 
 template< size_t M, typename T >
@@ -688,8 +661,6 @@ vector< M, T >::operator T*()
 {
     return array;
 }
-
-
 
 template< size_t M, typename T >
 vector< M, T >::operator const T*() const
@@ -712,9 +683,7 @@ vector< M, T >::operator[]( size_t index ) const
     return at( index );
 }
 
-
 #endif
-
 
 #if 0
 template< size_t M, typename T >
@@ -724,8 +693,6 @@ vector< M, T >::operator[]( size_t index )
     return at( index );
 }
 
-
-
 template< size_t M, typename T >
 inline const T&
 vector< M, T >::operator[]( size_t index ) const
@@ -733,7 +700,6 @@ vector< M, T >::operator[]( size_t index ) const
     return at( index );
 }
 #endif
-
 
 template< size_t M, typename T >
 vector< M, T >
@@ -745,8 +711,6 @@ vector< M, T >::operator*( const vector< M, T >& other ) const
     return result;
 }
 
-
-
 template< size_t M, typename T >
 vector< M, T >
 vector< M, T >::operator/( const vector< M, T >& other ) const
@@ -756,8 +720,6 @@ vector< M, T >::operator/( const vector< M, T >& other ) const
         result.at( index ) = at( index ) / other.at( index );
     return result;
 }
-
-
 
 template< size_t M, typename T >
 vector< M, T >
@@ -769,8 +731,6 @@ vector< M, T >::operator+( const vector< M, T >& other ) const
     return result;
 }
 
-
-
 template< size_t M, typename T >
 vector< M, T >
 vector< M, T >::operator-( const vector< M, T >& other ) const
@@ -781,9 +741,6 @@ vector< M, T >::operator-( const vector< M, T >& other ) const
     return result;
 }
 
-
-
-
 template< size_t M, typename T >
 void
 vector< M, T >::operator*=( const vector< M, T >& other )
@@ -791,8 +748,6 @@ vector< M, T >::operator*=( const vector< M, T >& other )
     for( size_t index = 0; index < M; ++index )
         at( index ) *= other.at( index );
 }
-
-
 
 template< size_t M, typename T >
 void
@@ -802,8 +757,6 @@ vector< M, T >::operator/=( const vector< M, T >& other )
         at( index ) /= other.at( index );
 }
 
-
-
 template< size_t M, typename T >
 void
 vector< M, T >::operator+=( const vector< M, T >& other )
@@ -812,8 +765,6 @@ vector< M, T >::operator+=( const vector< M, T >& other )
         at( index ) += other.at( index );
 }
 
-
-
 template< size_t M, typename T >
 void
 vector< M, T >::operator-=( const vector< M, T >& other )
@@ -821,8 +772,6 @@ vector< M, T >::operator-=( const vector< M, T >& other )
     for( size_t index = 0; index < M; ++index )
         at( index ) -= other.at( index );
 }
-
-
 
 template< size_t M, typename T >
 vector< M, T >
@@ -834,8 +783,6 @@ vector< M, T >::operator*( const T other ) const
     return result;
 }
 
-
-
 template< size_t M, typename T >
 vector< M, T >
 vector< M, T >::operator/( const T other ) const
@@ -845,8 +792,6 @@ vector< M, T >::operator/( const T other ) const
         result.at( index ) = at( index ) / other;
     return result;
 }
-
-
 
 template< size_t M, typename T >
 vector< M, T >
@@ -858,8 +803,6 @@ vector< M, T >::operator+( const T other ) const
     return result;
 }
 
-
-
 template< size_t M, typename T >
 vector< M, T >
 vector< M, T >::operator-( const T other ) const
@@ -870,9 +813,6 @@ vector< M, T >::operator-( const T other ) const
     return result;
 }
 
-
-
-
 template< size_t M, typename T >
 void
 vector< M, T >::operator*=( const T other )
@@ -880,8 +820,6 @@ vector< M, T >::operator*=( const T other )
     for( size_t index = 0; index < M; ++index )
         at( index ) *= other;
 }
-
-
 
 template< size_t M, typename T >
 void
@@ -891,8 +829,6 @@ vector< M, T >::operator/=( const T other )
         at( index ) /= other;
 }
 
-
-
 template< size_t M, typename T >
 void
 vector< M, T >::operator+=( const T other )
@@ -900,8 +836,6 @@ vector< M, T >::operator+=( const T other )
     for( size_t index = 0; index < M; ++index )
         at( index ) += other;
 }
-
-
 
 template< size_t M, typename T >
 void
@@ -911,8 +845,6 @@ vector< M, T >::operator-=( const T other )
         at( index ) -= other;
 }
 
-
-
 template< size_t M, typename T >
 vector< M, T >
 vector< M, T >::operator-() const
@@ -920,8 +852,6 @@ vector< M, T >::operator-() const
     vector< M, T > v( *this );
     return v.negate();
 }
-
-
 
 template< size_t M, typename T >
 const vector< M, T >&
@@ -932,16 +862,12 @@ vector< M, T >::negate()
     return *this;
 }
 
-
-
 template< size_t M, typename T >
 inline T&
 vector< M, T >::x()
 {
     return array[ 0 ];
 }
-
-
 
 template< size_t M, typename T >
 inline T&
@@ -950,16 +876,12 @@ vector< M, T >::y()
     return array[ 1 ];
 }
 
-
-
 template< size_t M, typename T >
 inline T&
 vector< M, T >::z()
 {
     return array[ 2 ];
 }
-
-
 
 template< size_t M, typename T >
 inline T&
@@ -968,16 +890,12 @@ vector< M, T >::w()
     return array[ 3 ];
 }
 
-
-
 template< size_t M, typename T >
 inline const T&
 vector< M, T >::x() const
 {
     return array[ 0 ];
 }
-
-
 
 template< size_t M, typename T >
 inline const T&
@@ -986,16 +904,12 @@ vector< M, T >::y() const
     return array[ 1 ];
 }
 
-
-
 template< size_t M, typename T >
 inline const T&
 vector< M, T >::z() const
 {
     return array[ 2 ];
 }
-
-
 
 template< size_t M, typename T >
 inline const T&
@@ -1004,15 +918,12 @@ vector< M, T >::w() const
     return array[ 3 ];
 }
 
-
 template< size_t M, typename T >
 inline T&
 vector< M, T >::r()
 {
     return array[ 0 ];
 }
-
-
 
 template< size_t M, typename T >
 inline T&
@@ -1021,16 +932,12 @@ vector< M, T >::g()
     return array[ 1 ];
 }
 
-
-
 template< size_t M, typename T >
 inline T&
 vector< M, T >::b()
 {
     return array[ 2 ];
 }
-
-
 
 template< size_t M, typename T >
 inline T&
@@ -1039,16 +946,12 @@ vector< M, T >::a()
     return array[ 3 ];
 }
 
-
-
 template< size_t M, typename T >
 inline const T&
 vector< M, T >::r() const
 {
     return array[ 0 ];
 }
-
-
 
 template< size_t M, typename T >
 inline const T&
@@ -1057,16 +960,12 @@ vector< M, T >::g() const
     return array[ 1 ];
 }
 
-
-
 template< size_t M, typename T >
 inline const T&
 vector< M, T >::b() const
 {
     return array[ 2 ];
 }
-
-
 
 template< size_t M, typename T >
 inline const T&
@@ -1075,32 +974,18 @@ vector< M, T >::a() const
     return array[ 3 ];
 }
 
-// result = vec1.cross( vec2 ) => result = vec1 x vec2
-template< size_t M, typename T >
-template< typename TT >
-inline vector< M, T > vector< M, T >::cross( const vector< M, TT >& rhs,
-                                 typename enable_if< M == 3, TT >::type* ) const
+template< size_t M, typename T > template< typename TT >
+vector< M, T >& vector< M, T >::cross( const vector< M, TT >& rhs,
+                                       typename enable_if< M == 3, TT >::type* )
 {
-    vector< M, T > result;
-    result.cross( *this, rhs );
-    return result;
+    const T x_ = y() * rhs.z() - z() * rhs.y();
+    const T y_ = z() * rhs.x() - x() * rhs.z();
+    const T z_ = x() * rhs.y() - y() * rhs.x();
+    x() = x_;
+    y() = y_;
+    z() = z_;
+    return *this;
 }
-
-
-
-// result.cross( vec1, vec2 ) => (this) = vec1 x vec2
-template< size_t M, typename T >
-template< typename TT >
-void vector< M, T >::cross( const vector< M, TT >& aa,
-                            const vector< M, TT >& bb,
-                            typename enable_if< M == 3, TT >::type* )
-{
-    array[ 0 ] = aa.y() * bb.z() - aa.z() * bb.y();
-    array[ 1 ] = aa.z() * bb.x() - aa.x() * bb.z();
-    array[ 2 ] = aa.x() * bb.y() - aa.y() * bb.x();
-}
-
-
 
 template< size_t M, typename T >
 inline T vector< M, T >::dot( const vector< M, T >& other ) const
@@ -1111,7 +996,6 @@ inline T vector< M, T >::dot( const vector< M, T >& other ) const
 
     return tmp;
 }
-
 
 template< size_t M, typename T >
 inline T vector< M, T >::normalize()
@@ -1142,16 +1026,12 @@ inline T vector< M, T >::squared_length() const
     return _squared_length;
 }
 
-
-
 template< size_t M, typename T >
 inline T
 vector< M, T >::distance( const vector< M, T >& other ) const
 {
     return std::sqrt( squared_distance( other ) );
 }
-
-
 
 template< size_t M, typename T >
 inline T vector< M, T >::squared_distance( const vector< M, T >& other ) const
@@ -1169,127 +1049,84 @@ template< size_t M, typename T > inline T vector< M, T >::product() const
     return result;
 }
 
-template< size_t M, typename T >
-void vector< M, T >::compute_normal( const vector< M, T >& aa,
-                                     const vector< M, T >& bb,
-                                     const vector< M, T >& cc )
+template< size_t M, typename T > template< typename TT >
+vector< M, T >& vector< M, T >::rotate( const T theta, vector< M, TT > axis,
+                                        typename enable_if< M == 3, TT >::type* )
 {
-    vector< M, T > u,v;
-    // right hand system, CCW triangle
-    u = bb - aa;
-    v = cc - aa;
-    cross( u, v );
-    normalize();
-}
-
-
-
-template< size_t M, typename T >
-vector< M, T >
-vector< M, T >::compute_normal( const vector< M, T >& bb,
-                                const vector< M, T >& cc ) const
-{
-    vector< M, T > tmp;
-    tmp.compute_normal( *this, bb, cc);
-    return tmp;
-}
-
-template< size_t M, typename T >
-template< typename TT >
-vector< 3, T > vector< M, T >::rotate( const T theta, vector< M, TT > axis,
-            typename enable_if< M == 3, TT >::type* ) const
-{
-    axis.normalize();
-
     const T costheta = std::cos( theta );
     const T sintheta = std::sin( theta );
+    axis.normalize();
 
-    return vector< 3, T >(
+    array[0] =
         (costheta + ( 1.0f - costheta ) * axis.x() * axis.x() ) * x()    +
         (( 1 - costheta ) * axis.x() * axis.y() - axis.z() * sintheta ) * y() +
-        (( 1 - costheta ) * axis.x() * axis.z() + axis.y() * sintheta ) * z(),
-
+        (( 1 - costheta ) * axis.x() * axis.z() + axis.y() * sintheta ) * z();
+    array[1] =
         (( 1 - costheta ) * axis.x() * axis.y() + axis.z() * sintheta ) * x() +
         ( costheta + ( 1 - costheta ) * axis.y() * axis.y() ) * y() +
-        (( 1 - costheta ) * axis.y() * axis.z() - axis.x() * sintheta ) * z(),
-
+        (( 1 - costheta ) * axis.y() * axis.z() - axis.x() * sintheta ) * z();
+    array[2] =
         (( 1 - costheta ) * axis.x() * axis.z() - axis.y() * sintheta ) * x() +
         (( 1 - costheta ) * axis.y() * axis.z() + axis.x() * sintheta ) * y() +
-        ( costheta + ( 1 - costheta ) * axis.z() * axis.z() ) * z() );
+        ( costheta + ( 1 - costheta ) * axis.z() * axis.z() ) * z();
+    return *this;
 }
 
-
 // sphere layout: center xyz, radius w
-template< size_t M, typename T >
-template< typename TT >
-inline vector< 3, T >
-vector< M, T >::
-project_point_onto_sphere( const vector< 3, TT >& point,
+template< size_t M, typename T > template< typename TT > inline vector< 3, T >
+vector< M, T >::project_point_onto_sphere( const vector< 3, TT >& point,
     typename enable_if< M == 4, TT >::type* ) const
 {
-    const vector< 3, T >& _center = get_sub_vector< 3 >( 0 );
+    const vector< 3, T >& center_ = get_sub_vector< 3 >();
 
     vector< 3, T > projected_point( point );
-    projected_point -= _center;
+    projected_point -= center_;
     projected_point.normalize();
     projected_point *= w();
-    return _center + projected_point;
+    return center_ + projected_point;
 }
 
-
-
 // sphere layout: center xyz, radius w
-template< size_t M, typename T >
-template< typename TT >
-inline T
-vector< M, T >::
-distance_to_sphere( const vector< 3, TT >& point,
-    typename enable_if< M == 4, TT >::type* ) const
+template< size_t M, typename T > template< typename TT > inline T
+vector< M, T >::distance_to_sphere( const vector< 3, TT >& point,
+                                    typename enable_if< M == 4, TT >::type* )
+    const
 {
-    const vector< 3, T >& center_ = get_sub_vector< 3 >( 0 );
+    const vector< 3, T >& center_ = get_sub_vector< 3 >();
     return ( point - center_ ).length() - w();
 }
 
-template< size_t M, typename T > template< size_t N > inline
-vector< N, T >& vector< M, T >::get_sub_vector( size_t offset,
-                                           typename enable_if< M >= N >::type* )
+template< size_t M, typename T > template< size_t N, size_t O >
+vector< N, T > vector< M, T >::get_sub_vector(
+    typename enable_if< M >= N+O >::type* ) const
 {
-    assert( offset <= M - N );
-    return reinterpret_cast< vector< N, T >& >( *( begin() + offset ) );
+    return vector< N, T >( array + O );
 }
 
-template< size_t M, typename T > template< size_t N > inline
-const vector< N, T >& vector< M, T >::get_sub_vector( size_t offset,
-                                     typename enable_if< M >= N >::type* ) const
+template< size_t M, typename T > template< size_t N, size_t O >
+void vector< M, T >::set_sub_vector( const vector< N, T >& sub,
+                                     typename enable_if< M >= N+O >::type* )
 {
-    assert( offset <= M - N );
-    return reinterpret_cast< const vector< N, T >& >( *( begin() + offset ) );
+    ::memcpy( array + O, sub.array, N * sizeof( T ));
 }
-
 
 // plane: normal xyz, distance w
 template< size_t M, typename T > template< typename TT >
 inline T vector< M, T >::distance_to_plane( const vector< 3, TT >& point,
     typename enable_if< M == 4, TT >::type* ) const
 {
-    const vector< 3, T >& normal = get_sub_vector< 3 >( 0 );
+    const vector< 3, T >& normal = get_sub_vector< 3 >();
     return normal.dot( point ) + w();
 }
 
-
-
 // plane: normal xyz, distance w
-template< size_t M, typename T >
-template< typename TT >
-vector< 3, T >
+template< size_t M, typename T > template< typename TT > vector< 3, T >
 vector< M, T >::project_point_onto_plane( const vector< 3, TT >& point,
     typename enable_if< M == 4, TT >::type* ) const
 {
-    const vector< 3, T >& normal = get_sub_vector< 3 >( 0 );
+    const vector< 3, T >& normal = get_sub_vector< 3 >();
     return point - ( normal * distance_to_plane( point ) );
 }
-
-
 
 template< size_t M, typename T >
 bool vector< M, T >::operator==( const vector< M, T >& other ) const
@@ -1297,13 +1134,11 @@ bool vector< M, T >::operator==( const vector< M, T >& other ) const
     return memcmp( array, other.array, sizeof( array )) == 0;
 }
 
-
 template< size_t M, typename T >
 bool vector< M, T >::operator!=( const vector< M, T >& other ) const
 {
     return ! this->operator==( other );
 }
-
 
 template< size_t M, typename T >
 bool
@@ -1315,7 +1150,6 @@ vector< M, T >::equals( const vector< M, T >& other, T tolerance ) const
     return true;
 
 }
-
 
 template< size_t M, typename T >
 bool
@@ -1329,7 +1163,6 @@ vector< M, T >::operator<( const vector< M, T >& other ) const
     return false;
 }
 
-
 // to-homogenous-coordinates assignment operator
 // non-chainable because of sfinae
 template< size_t M, typename T > template< size_t N >
@@ -1340,7 +1173,6 @@ vector< M, T >::operator=( const vector< N, T >& source_ )
     at( M - 1 ) = static_cast< T >( 1.0 );
     return 0;
 }
-
 
 // from-homogenous-coordinates assignment operator
 // non-chainable because of sfinae
@@ -1355,15 +1187,12 @@ vector< M, T >::operator=( const vector< N, T >& source_ )
     return 0;
 }
 
-
 template< size_t M, typename T >
 vector< M, T >& vector< M, T >::operator=( const T* c_array )
 {
     iter_set( c_array, c_array + M );
     return *this;
 }
-
-
 
 template< size_t M, typename T >
 T vector< M, T >::operator=( T filler_value )
@@ -1373,9 +1202,6 @@ T vector< M, T >::operator=( T filler_value )
     return filler_value;
 }
 
-
-
-
 template< size_t M, typename T >
 vector< M, T >& vector< M, T >::operator=( const vector< M, T >& other )
 {
@@ -1383,8 +1209,6 @@ vector< M, T >& vector< M, T >::operator=( const vector< M, T >& other )
         memcpy( array, other.array, M * sizeof( T ) );
     return *this;
 }
-
-
 
 // returns void to avoid 'silent' loss of precision when chaining
 template< size_t M, typename T > template< typename U >
@@ -1395,8 +1219,6 @@ void vector< M, T >::operator=( const vector< M, U >& source_ )
     for( iterator my_it = begin(); it != it_end; ++it, ++my_it )
         *my_it = static_cast< T >( *it );
 }
-
-
 
 template< size_t M, typename T >
 template< typename input_iterator_t >
@@ -1421,36 +1243,12 @@ void vector< M, T >::clamp( const T& min, const T& max )
     }
 }
 
-
-
-template< size_t M, typename T >
-template< typename TT >
-void
-vector< M, T >::scale_to( vector< M, TT >& result_,
-    T min_value, T max_value ) const
-{
-    T range       = max_value-min_value;
-    T half_range  = range * 0.5;
-    T scale       = ( 1.0 / range ) * static_cast< T >( std::numeric_limits< TT >::max() );
-
-    for( size_t index = 0; index < M; ++index )
-    {
-        result_.at( index )
-            = static_cast< TT >( ( at( index ) + half_range ) * scale );
-    }
-
-}
-
-
-
 template< size_t M, typename T >
 inline size_t
 vector< M, T >::size()
 {
     return M;
 }
-
-
 
 template< size_t M, typename T >
 size_t
@@ -1459,16 +1257,12 @@ vector< M, T >::find_min_index() const
     return std::min_element( begin(), end() ) - begin();
 }
 
-
-
 template< size_t M, typename T >
 size_t
 vector< M, T >::find_max_index() const
 {
     return std::max_element( begin(), end() ) - begin();
 }
-
-
 
 template< size_t M, typename T >
 size_t
@@ -1477,16 +1271,12 @@ vector< M, T >::find_abs_min_index() const
     return std::min_element( begin(), end(), vmml::math::abs_less< T >() ) - begin();
 }
 
-
-
 template< size_t M, typename T >
 size_t
 vector< M, T >::find_abs_max_index() const
 {
     return std::max_element( begin(), end(), vmml::math::abs_greater< T >() ) - begin();
 }
-
-
 
 template< size_t M, typename T >
 T&
@@ -1495,16 +1285,12 @@ vector< M, T >::find_min()
     return *std::min_element( begin(), end() );
 }
 
-
-
 template< size_t M, typename T >
 const T&
 vector< M, T >::find_min() const
 {
     return *std::min_element( begin(), end() );
 }
-
-
 
 template< size_t M, typename T >
 T&
@@ -1513,15 +1299,12 @@ vector< M, T >::find_max()
     return *std::max_element( begin(), end() );
 }
 
-
-
 template< size_t M, typename T >
 const T&
 vector< M, T >::find_max() const
 {
     return *std::max_element( begin(), end() );
 }
-
 
 template< size_t M, typename T >
 inline typename vector< M, T >::iterator
@@ -1530,14 +1313,12 @@ vector< M, T >::begin()
     return array;
 }
 
-
 template< size_t M, typename T >
 inline typename vector< M, T >::iterator
 vector< M, T >::end()
 {
     return array + M; ;
 }
-
 
 template< size_t M, typename T >
 inline typename vector< M, T >::const_iterator
@@ -1546,15 +1327,12 @@ vector< M, T >::begin() const
     return array;
 }
 
-
 template< size_t M, typename T >
 inline typename vector< M, T >::const_iterator
 vector< M, T >::end() const
 {
     return array + M; ;
 }
-
-
 
 template< size_t M, typename T >
 inline typename vector< M, T >::reverse_iterator
@@ -1563,14 +1341,12 @@ vector< M, T >::rbegin()
     return array + M - 1;
 }
 
-
 template< size_t M, typename T >
 inline typename vector< M, T >::reverse_iterator
 vector< M, T >::rend()
 {
     return array - 1;
 }
-
 
 template< size_t M, typename T >
 inline typename vector< M, T >::const_reverse_iterator
@@ -1579,15 +1355,12 @@ vector< M, T >::rbegin() const
     return array + M - 1;
 }
 
-
 template< size_t M, typename T >
 inline typename vector< M, T >::const_reverse_iterator
 vector< M, T >::rend() const
 {
     return array - 1;
 }
-
-
 
 template< size_t M, typename T >
 bool
@@ -1611,9 +1384,6 @@ vector< M, T >::is_unit_vector() const
     return one;
 }
 
-
-
-
 template< size_t M, typename T >
 void
 vector< M, T >::perturb( T perturbation )
@@ -1635,8 +1405,6 @@ vector< M, T >::sqrt_elementwise()
     }
 }
 
-
-
 template< size_t M, typename T >
 void
 vector< M, T >::reciprocal()
@@ -1646,8 +1414,6 @@ vector< M, T >::reciprocal()
         (*it) = static_cast< T >( 1.0 ) / (*it);
     }
 }
-
-
 
 template< size_t M, typename T >
 void
@@ -1663,8 +1429,6 @@ vector< M, T >::reciprocal_safe()
             v = static_cast< T >( 1.0 ) / v;
     }
 }
-
-
 
 template< size_t M, typename T >
 template< typename TT >
@@ -1700,7 +1464,6 @@ vector< M, T >::nnz() const
     return counter;
 }
 
-
 template< size_t M, typename T >
 double
 vector< M, T >::norm( ) const
@@ -1729,7 +1492,6 @@ vector< M, T >::set_random( int seed )
         at( i ) = -1.0 + 2.0 * fillValue;
     }
 }
-
 
 } // namespace vmml